requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.49
botocore==1.34.49
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.55.2
django-debug-toolbar==4.1.0
django-storages==1.14.2
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==7.6.0
typing_extensions==4.7.1
urllib3==1.26.18
